Add proper SHA-3
Kind of a copy and paste of Keccak, but only a single copy of the
permutation at least. Keccak depends on SHA-3 instead of the reverse,
so that SHA-3 can be enabled without also bringing in an unapproved
hash function.

Updates newhope code and removes API function newhope_hash which was
an unofficial SHA-3-256.