Tighten up TLS server handshake logic.
Previously client was allowed to omit the Certificate message, a
leftover from supporting SSLv3. In all versions of TLS, an empty
message must be sent if the client does not want to use a cert.
No known security impact, but nothing we need to allow anymore.

Clean up the handshake switch a bit by using return statements.