Expose pk_ops.h as public interface again
I was initially thinking of Botan as somewhat closed system, but
@cordney has a legit use case of wanting to expose a TPM sign
operation, but using their internal TPM library and TPM key type.
This requires the API be exposed so the derivation can occur.

Add a comment in the header explaining that the header is not for
normal application use and directing them to pubkey.h

This basically reverts 2747e8e23aec43162