Update shared object naming for new versioning scheme.
Cleans up so object naming since most of the time (across Unix)
we follow the exact same naming scheme; just make it the default
if only the so suffix is specified in the file.

Also updates include header dir to be botan-${major}

Changes behavior when shared lib not supported; instead of making
the user explicitly try again with --disable-shared, just assume it
and continue running.