Add botan_ffi_supports_api function
This lets us upgrade the FFI version over time and still allow
applications to reliably detect if the current library binary
supports their version.

As an example, it would be useful to be able to add features to FFI
sometime in 2.x. In that case, we would increase the value of the
FFI API version, even though anything calling the old API would still
work perfectly. Applications can verify at runtime the API they want to
use is supported using this new call.