Horrible hack for OS X
On Travis OS X, platform.machine() is 'x86_64' and
platform.processor() is 'i386'. We incorrectly choose a 32-bit
build, which fails due to no 32-bit Boost libraries.

Switch (fairly arbitrary) order of machine and processor for CI.
I have a plan to address this better but it will be for 2.1