Load pkcs11 module by request
The pkcs11 module once required the pkcs11 headers
as an external dependency, but the headers were included
a while ago. Still, the module was set to be load_on vendor.
Changes it to be loaded by request. This seems the right
choice, given that not all users would need pkcs11 support.