configure: Remove duplicates from generated module list
I ran into this after PKCS11 was enabled by default, as my local
build script uses --enable-modules=pkcs11,... this ended up causing
the module to be loaded twice! The result was duplicate entries in
the Makefile.

Would be good for configure to be written more defensively, the
result of this error being a bogus Makefile is lame.