Updates to the readme
Instead of refering to security.rst for how to report a security bug,
just say so in the readme.

Mention the botan-announce list.

Drop mention of gitter.im as place for reporting issues, it is not
good for that. GitHub issues and mailing list are both part of our
infrastructure and have a large audience, gitter.im is kind of
peripheral and not monitored so much.

[ci skip]