Add CPUID::to_string
Deprecates CPUID::print, mainly to avoid use of ostream in utils.

Main reasoning is some platforms do not have streams at all. If all uses of
streams are for operator<< and operator>> overloads, these can be guarded by a
future BOTAN_HAS_CPP_STREAMS system-feature flag.  But when ostreams are
required to use the API (as here), that is not possible.