Add toggle to disable SIMD speed test added in 2a7127fea6
Failed to consider need for ISA flags (none required on x86-64 or
aarch64 which are my primary test targets) but handling this correctly
requires splitting up speed.cpp which requires a lot more refactoring
work than I want to take on. Anyway the additional complexity just for
this microbenchmark is not worth it so hide it behind a macro check.

Also add an ECC scalar perf test.