The certificate being tested by the OCSP online test has expired.
This caused a test failure if run with --run-online-tests. This is not
the default but it caused the coverage build to fail.

Fix the test to reflect expired status. The whole thing does not
seem wholly satsifactory since we are still relying the behavior on an
external party, perhaps a mock OCSP responder should be created for
the tests.