TLS client must not pass an IP address as server name
RFC 6066 section 3 says: Literal IPv4 and IPv6 addresses are not
permitted in "HostName".  But if a user passes an IP address to
botan tls_client as connect address, this is also used for SNI.
Some TLS server like libtls from the LibreSSL project check that a
provided hostname is a DNS name.  The TLS connection attempt from
botan is rejected with a fatal alert.

Fix the issue in the botan library.  Do not send a SNI extension
in the client hello if the hostname is empty or an IPv4 or IPv6
address instead of a DNS name.