Small ffi fixes
Constify arg to botan_pk_op_decrypt, and correct comment on
botan_mp_is_positive, which returns true even for zero.

Considered changing botan_mp_is_positive to match the comment, and
return false for zero, but we already test that botan_mp_is_positive(0) == 1,
so I left it as is and fixed the comment.