Allow ECDSA tests for OpenSSL by comparing EC Groups
The ecdsa_sign test does not work for OpenSSL since the public key
that is derived from the private key does not store the OID of the
curve.  Thus, the OpenSSL code cannot find the proper NID for the
given public key and cannot do a verification operation.  We can
fix this by comparing the actual groups and not using the OIDs.