BearSSL: Initial support and hash tests
BearSSL is an implementation of the SSL/TLS protocol in C aiming
to be correct and secure, small and highly portable.  Thus making
it nicer to be included in a rather sparse bootloader.  This commit
adds support for BearSSL's hash routines only, with more stuff
coming up in following commits.  The goal is to be able to test
BearSSL using Botan's extensive testsuite.