Remove "pkcs11" provider paramater for pk ops in handbook and tests
We have special key types for PKCS#11 public and private keys.
When using the PK ops classes, passing the "pkcs11" provider
parameter is not necessary, as the PK ops is internally chosen
by the PKCS#11 key class, in contrast to other providers such
as openssl, which don't have special key classes. Updates the
handbook code examples and the tests.