Turn --destdir option into an environment variable, fixes #1101, fixes #996, supersedes PR #997
This commit removes the (broken) --destdir command line option and introduces the
DESTDIR environment variable for the 'make install' target. When it's set, the installation will
take place in the DESTDIR directory, under the prefix set by --prefix, with all of the internal
references (pkgconfig file, etc) containing only the prefix.
This behavior is more standard (as in autotools and co.), so makes packaging easier.