Change how CPU endianness is detected.
This is kind of an artificial change to appease Coverity but maybe
a bit cleaner in any case.

The issue is Coverity detects the throw in the case of PDP endian
being found at runtime, and assumes it might happen. And since
OS::get_processor_timestamp calls CPUID on x86 (to check for RDTSC),
it might throw. And so on up the call chain until it detect a throw
escaping from ~Timer_Scope in speed.cpp, which would crash.

However I do not have a PDP-endian x86 around to confirm this ... :)