Add guards for MAC and hash algo used in  passhash9 tests
Previously, passhash9 tests would fail if for example blowfish was not part
of the build. Adds guards for the different MAC and hash algorithms
used in passhash9 for calls to generate_passhash9(). For check_passhash9(),
there is no way to know that an algorithm is not supported, but at least
we improve the situation a bit here.