Some attempts at speeding up CI
First do amalgamation build on static instead of shared build. This
is arbitrary, except that we do 3 shared builds, and only one static
lib build, and the amalgamation kind of screws up ccache.

Also avoid installing LZMA or Python3 on OS X. Just running Homebrew
to install turns out to be significantly more expensive than either
building the library or running the tests (!).

Removes the workaround for homebrew #42553 because apparently it
is fixed now.