Simplify polynomial doubling code
GCC and Clang generate effectively identical code for a template
with parameters, vs completely unrolled code as was used previously.

Add a little-endian variant so XTS can use it. This extends XTS support
to cover 256 and 512-bit ciphers. I was not able to find another
implementation that supports both XTS and ciphers with large blocks,
so the XTS test vectors are self-generated.