Revert SIMDing of xor_buf in f39aa7bb
Forgot this causes all kinds of issues when the SIMD ISA is not
part of the base instruction set (needing -maltivec, etc). It just
happens to work on x86-64 because SSE2 is part of the base ISA there.

Instead just unroll by 16x this seems to be enough that GCC figures it
out.