Better tests for SIV
Correct errors in the AEAD tests that assumed process/update always
return something - that isn't true for SIV

Minor optimizations in CMAC and CTR to cache the block size instead
of making a zillion virtual calls for it.

Generalize SIV slightly to where it could support a non-128 bit
cipher, but don't pull the trigger on it since I can't find any
implementations to crosscheck with.