Use a simple PRNG for the tests
Not cryptographically secure, but fast! Cuts several seconds off
the test suite even on a very fast machine. Probably even more
effective for 32-bit systems since the default for HMAC_DRBG is
SHA-384. Also it means deterministic tests are used regardless
of build configuration which is nice.

Improve output for --test-runs which was useful for me when
debugging SM2 encryption issue.