Remove unaligned mem check
This previously enabled doing something unsafe (misaligned reads),
but it turns out even on hardware that supports this, it is not
safe to do because the compiler may do something unfortunate.
Now memcpy is used, which is safe on any platform.

Should provide a noticable speedup for ARM and PPC64, which previously
used the byte-at-a-time fallback code.