Fix a bug in PEM decoding
If the label param was set to a non-empty string, the decoding would treat
it as part of the label and then end up throwing a decoding error. The key
fix here is adding the clear() call in decode. Private bug report by email.

Add some basic tests for the PEM decoder.