Build improvements
Derive the fact that shared libs are not supported from the lack of
information about shared object names. If soname_suffix and co are
not set -> shared libs are disabled.

Disable stack protector by default on OSes where it has proven a problem
namely AIX and Haiku. Add --with-stack-protector option to force turning
it on.