Improve output of the ASN1 printer
The output was pretty much wrong for application-tagged types. Instead
the type was printed as if it was a universal tag.

Add a hack in the ASN1 printer for GeneralNames. These are a commonly
used implicitly tagged type. Basically if it is a context specific field,
and it looks like it might be a printable URI or DNS GeneralName, treat it
as a string.