Avoid using monty workspace for reduce_below
If the workspace is swapped, then it is too small for the Montgomery
operation and will be reallocate on the next sqr/multiply operation.

Also use ws[9] consistently for the Montgomery workspace, otherwise
if add needs to pass off the mult2, the workspaces are not the expected
size and again a reallocation occurs.