Avoid using threads in DSA operations
For DSA signing using a thread turned out to be purely a pessimization.
The single threaded code is faster even on a 4-core machine running Linux
(which has very fast thread creation). It would likely be much worse
on a single core machine or an OS with slower thread primitives.

For DSA verification, use Montgomery multi-exponentiation instead.