Remove use of rpath=ORIGIN and similar tricks on Darwin
These complicated the build steps, and distros don't like (or
explicitly prohibit) use of rpath=ORIGIN for security reasons.

The only reason for doing this was so that the user would not
be confused, by trying to run botan-test without their LD_LIBRARY_PATH
pointing in the right spot. But it is easier to just fix that
problem by adding a script that sets up the environment and then
runs the binary.