Fixes XMSS leaf index bounds sanity check
Prior to this patch the sanity check for XMSS leaf indices was wrongly based on
the tree height. As a result only half of the one-time keys could be used.

Instead base leaf index sanity check on the number of levels in a tree which
equals tree height + 1.

(see also: https://tools.ietf.org/html/draft-irtf-cfrg-xmss-hash-based-signatures-12#section-4.1.1)