Remove build time toggle for ECC coordinate masking
This is not a decision we should leave to end users.

And always use a random mask equal in size to the underlying field.
It was never quite clear if 80 bits was sufficient or not. But
taking a random field element is clearly the best possible situation,
and has very little additional cost.