Avoid useless multiplication in Montgomery exponentiation
When beginning the loop we initialized a value to one (in Montgomery form)
then multiply it by the first element looked up based on the exponent.
But this will always (after Montgomery multiplication) be exactly the
value we looked up in the table. So just assign it directly and avoid
the redundant operation.

Improves RSA verification by 5% or so since the number of multiplications
is so small in that case saving even 1 in useful. For other operations
there is no measurable improvement.