Use a smaller sieve when generating primes
This was the original behavior but 5af44a91ad switched the sieve
to always be the size of the hardcoded prime table. But this ends
up being quite a bit slower than necessary. Instead use as many
sieve elements as bits in the desired prime which is probably not
precisely optimal but seems to provide good speedups for both 1024
and 2048 bit prime generation. This is especially notable when generating
strong primes.