Add ChaCha using SIMD_4x32
This allows supporting SSE2, NEON and AltiVec in a single codebase,
so drop the NEON and SSE2 code.

This new impl avoids having to do shuffles with every round and so
is about 10% faster on Skylake.

Also, fix bugs in both baseline and AVX2 implementations when the
low counter overflowed. The SSE2 and NEON code were also buggy here.