Disable Sonar build
It is suddenly 5x slower for an unknown reason
https://community.sonarsource.com/t/sonar-analysis-step-suddenly-much-5x-slower-on-travis-ci/4364

Now it either times out or else completely dominates the runtime for Travis,
either is not acceptable.

Leave the framework for using Sonar in CI for now so it is easy to test
if things have improved later. If things are still broken in a month or
two I'll probably just drop support for Sonar.