Increase Travis git pull depth
Undocumented? side effect of a small git pull depth - if more than N
new commits are pushed to master while an earlier build is running,
the old build starts failing, as when CI does the pull it does not
find the commit it is building within the checked out tree.