Add a multi-file input mode for test fuzzers
The test_fuzzers.py script is very slow especially on CI. Add a mode
to the test fuzzers where it will accept many files on the command
line and test each of them in turn. This is 100s of times faster,
as it avoids all overhead from fork/exec.

It has the downside that you can't tell which input caused a crash, so
retain the old mode with --one-at-a-time option for debugging work.