Use TLS::Context::verifyCallback as tls_verify_cert_chain callback
StreamCore uses the user-provided verify callback as a
tls_verify_cert_chain if any is provided to the Context. Stream allows
configuring the context as well using Stream::set_verify_callback.
Stream now keeps a reference to the Context, rather than copying it.
This allows users to configure the Context after constructing the
Stream, but requires them to manage its lifetime.