Officially deprecate headers
Create BOTAN_DEPRECATED_HEADER so we can warn about this consistently.

Shuffle around the filter headers so all of the concrete filters
are defined in filters.h instead of being spread across many headers.

Document which headers are deprecated as well as a list of headers which
will be made internal-only in a future major release.