Only use getauxval replacement on ARMv7
Since this code only works on 32-bit systems due to casts to 32-bit ELF specific
types. The code should be completely unnecessary on Aarch64 since the oldest
supported Android API version for 64-bit supports getauxval, but it is possible
someone not knowing that might disable use of native getauxval for a 64-bit build.