Remove references to Kullo CI
With the increasing stability of Botan and the use of other CI services,
this CI became obsolete over the years. Now that there is an Android
build on Travis, Kullo CI does not add sufficient value anymore to
justify the maintenance work.

It has been a pleasure contributing to the development of Botan, helping
to bring the project to the maturity it has now.