Have Clang print the deprecated message
I cannot figure out exactly when Clang started supporting a message on the
deprecated attribute. It is documented as supporting such for 3.9. But in 3.8.0
and all earlier versions of the reference, the deprecated attribute isn't
mentioned at all.

Just enable it for all versions. If there happens to be a version of Clang which
both doesn't support messages on deprecated attributes but can otherwise
sucessfully compile the library, someone will probably tell us about it.