Fix some dangling pointer problems
After running Clang's experimental -Wlifetime over the Botan codebase,
I discovered a few dangling pointer problems that resulted from
returning a pointer to a temporary object. The temporary object will
be destroyed at the end of the scope and the pointer will dangle.

Instead of returning a dangling pointer, return a `std::unique_ptr` to
make more explicit that it's the responsibility of the caller to
destroy the resource that was handled in by the function. In some
cases, like `subject_public_key()`, it was explicitly said in
comments ("This object is owned by the caller."), but returning a
explicit type that transfer ownership (a unique pointer) is better.