Remove support for splitting the amalgamation by ABI
This is as if --single-file-amalgamation was always used, except
also now botan_all_internal.h is not created.

This effectively drops support for very old GCC/Clang in the
amalgamation (only). GCC 5+ and Clang 3.8+ support the target
attribute and work fine. MSVC not affected since it doesn't need
such attributes in the first place.