Increase size of ECC scalar blinding
Discussion in #880 and #893

I can't remember when or why this switched to fixed 80 bit mask
but choosing based on the size of the order is better especially
for groups with an order very close to a power of 2 like P-384.

Some performance loss but seems tolerable; with all the recent
optimization work, ECDSA P-256 with this change is still faster than
in 2.13.0. Larger groups like P-384/P-521 take more of a hit.